feat: Delay yielding const/loadConst commands (#153)
Only output them immediately before they are consumed by an operation.

This ensures that constant parameters are closer to their consumers on
the traversal.

That's specially useful for `CircuitChunks`, so we can avoid loading a
const in one chunk that's only used another one.
(Before this change, the first chunk always had all the `LoadConst`
nodes).