feat: Multithreading improvements (#182)
The main feature of this PR is reducing the memory usage of
multithreading with a priority channel by adding a shared
`Arc<RwLock<Option<P>>>` with the maximum circuit cost in the queue, so
the workers don't fill the channels with useless data.

The main improvement of this is limiting the memory usage when multiple
threads are involved. Now I can run `barenco_tof_10` on 10 threads with
~4GB of RAM, where before it exceeded 10GB (and caused a OOM).

I also included multiple small changes extracted from #175;

- `CircuitCost::as_usize` instead of proxy operations, so we can share
cost limits as atomics and run other primitive operations directly.
- Print the actual number of circuits processed on the log (not only the
seen count, which includes elements in the queue).
- Replace `eprintln!`s with log calls.
- Print the max queue capacity at the beginning of the execution.