feat: Exhaustive greedy rewrite strategy (#151)
This is a straight combination of the greedy and exhaustive strategies.
Instead of applying each strategy separately, apply each one once but
also include all the following strategies that do not modify the same
nodes.

This is a draft, since for some reason combining multiple strategies at
once is invalidating the hugr.
They may not be correctly reporting their modified nodes ?