feat: Simplify the circ hash code. Allow nodes with children. (#202)
Before this change, we stored the hash for each port in a map. We could
then drop it when consumed if the port was linear.
This wasn't really necessary; the memory usage reduction vs just storing
all the node hashes is quite small in comparison with the circuit size,
and the extra logic made things slower.

This PR replaces that logic with a straight "compute each node hash and
store it". This is more robust in cases that have non-local edges.

I also added a recursive call to support hashing dataflow container
nodes.