refactor: Move python bindings from `tket2` to `tket2-py` (#235)
Adds transparent wrappers on `tket2-py` that derive pyclass themselves.
This has multiple goals:

- Cleans up the code. Python stuff goes into the python crate.
Now `tket2:pyo3` and `hugr:pyo3` are only used for the python error
implementations.
(We could potentially move this too, and avoid us the `pyo3` dependency
headache).
  
- Let's us use the `tket2-py` conversion methods for tket1 and tket2
circuits. This is required for #178, which is in turn required for all
the python bindings in #204 that require stable indices.

Closes #148 

---

The code is mostly the same, asides from:
- replacing the custom implementation of `PyPatternMatch` with a
transparent wrapper. (There was a TODO for this).
- updating to tket-json-rs 0.3 (cleans up some calls that use
`Python::with_gil`)