fix: Check for rewrite composition in badger
This is a short-term quick-but-incomplete fix for the rewrite
composition generating loops in the circuit.

We now check right after the composition whether we have invalidated the
circuit.

Still, it is theoretically possible for an invalid chain of rewrites to
end up producing a valid (but not equivalent) circuit.
There is a discussion open for discussing more general solutions to this problem.