fix: Check for rewrite composition in badger (#255)
This is a short-term quick-but-incomplete fix for #239.

We now check right after the composition whether we have invalidated the
circuit by making a loop.
We leverage that circuit hashing already does a toposort traversal,
adding an error when a loop is found and catching it.

Still, it is theoretically possible for an invalid chain of rewrites to
end up producing a valid (but not equivalent) circuit. There is a
discussion open for discussing more general solutions to this problem
#242.