feat: Implement `PyErr` conversion locally in `tket2-py` (#258)
With this we can drop the `pyo3` features in hugr and portgraph.
`tket2` and `tket-json-rs` still require it, as they define some
non-error pyclasses.

This requires manually converting errors by calling `convert_pyerrs`
before `?`, but it is a small price to pay for not dealing with pyo3
versions across crates.