feat!: Lexicographic cost fn (#277)
The major+minor circuit cost was generalised to an arbitrary
lexicographic cost in #270. The old code now corresponds to the special
case N=2.

This PR modifies the cost functions related to it to reflect that
generalisation. I've also taken the opportunity to change the language a
bit. I hope naming is now clearer and more consistent.

---------

Co-authored-by: Seyon Sivarajah <seyon.sivarajah@cambridgequantum.com>