chore(deps): update pyo3 dependency to 0.21 (#314)
This update has a lot of breaking changes, but most of the changes in
this PR are just related to replacing rust references with the
`Bound<'py, T>` smart pointer (which keeps track of the python
lifetimes).
See migration guide: https://pyo3.rs/v0.21.1/migration.html

Also updated `tket-json-rs` to `0.4.0`, to match the pyo3 dependency.