ci: Conditional checks based on changed files (#371)
Bring the conditional CI checks from hugr, so we don't run unnecessary
steps all the time and we can parallelise a bit more.

If the PR is labelled with `run-ci-checks`, it will run all the checks
independently of the changed files.

I'll update the repo `required` checks before merging this.