feat: `Circuit::operations` (#395)
Returns a subset of `Circuit::commands`, corresponding to the commands
counted by `Circuit::num_operations`.

This is useful to ignore `LoadConst`s and tuple operations when we only
care about "gate like" operations in the circuit. But mainly, it's meant
as an iterator with similar behaviour to `num_operations`.

Note that #199 is still an open issue; tuple operations in the middle of
the circuit may result in confusing linear indices.