fix: Support non-DFG circuits (#391)
Followup to #390.
Fixes #389.
Closes #385.
~~Blocked by https://github.com/CQCL/hugr/issues/1175.~~

Fixes support for non-`Dfg` circuits and circuits with a non-root
parent:

Adds a `Circuit::extract_dfg(&self)` function that extracts the circuit
into a new hugr with a DFG operation at the root.
In some cases, like in a CFG DataflowBlock node, this requires some
changes to the definition to eliminate the output sum type.
Here I only implemented it for the kind of blocks produced by guppy. We
could replace the manual implementation once
https://github.com/CQCL/hugr/issues/818 gets implemented.

With this we can now fix #389 by extracting the circuit before using it
as a replacement in `SimpleReplacement::create_simple_replacement`.

Replaces `DfgBuilder` with `FunctionBuilder` where possible, so we can
use named circuits in the tests.
(This failed before due to the bug in CircuitRewrite).