refactor: Cleanup tket1 serialized op structures (#419)
This is a noisy internal refactor of `::serialize::pytket::op::JsonOp`,
extracted from the work towards #379.

`JsonOp` was a temporary structure used during the encoding/decoding of
pytket circuits that represented two different kinds of operation:
- pytket operations with a direct tket2 counterpart
- other operations that have to be encoded as OpaqueOps

This mixed up the two definitions, and made applying custom logic to one
of the variants more annoying.
(E.g. the special handling of bit input/outputs for tket2 ops needed for
#379).

This PR splits the structs into a `Native` and an `Opaque` variant, so
we can keep the implementation clean. The code is functionally the same.

---------

Co-authored-by: doug-q <141026920+doug-q@users.noreply.github.com>