fix: `extract_dfg` inserting the output node with an invalid child order (#442)
When replacing a CFG's output with the correct DFG output operation, we
inserted the new node at the end of the container's children, instead of
just after the input.

Includes a new unit test which simulates a circuit definition from
guppy, where we can run the `lower_to_pytket` pass.