chore: Use local ruff and mypy for pre-commit checks (#547)
Ensures that the checks run by pre-commit are the same as the ones run
in CI, and the `ruff` and `mypy` versions are always controlled by
`pyproject.toml`.
Before, we used pre-commit actions with their own versions that needed
to be updated separately and tended to get out of date.