chore: Replace `poetry` with `uv` (#557)
Replaces the python project manager with `uv`.
For most commands that used poetry before, it just be as easy as
replacing `poetry {things}` with `uv {things}`.

Some notable exeptions:

- Instead of `poetry install --with extra1,extra2`,
  use `uv sync --extra extra1 --extra extra2`.
Note that will delete any extraneous packages/extras so the venv state
is deterministic.
  Add `--inexact` if you want to replicate the old poetry behaviour.

- `poetry shell` is no longer available. The preferred workflow is to
`uv run` / `uvx` things, to ensure you are always in sync with the
project definition.
If that's too annoying, it is always possible to `source
.venv/bin/activate`.

Updates the justfile, so using that should be the same as before.

---------

Co-authored-by: Seyon Sivarajah <seyon.sivarajah@quantinuum.com>