feat!: Update rust hugr dependency to `0.12.0`, and python hugr to `0.8.0` (#568)
The main change is the drop of `CustomOp`s, requiring opaque ops to
always be resolved.

This is a bit of a problem with the current code, as we used `OpaqueOp`s
not backed by a definition to declare operations from python.
For the moment, we leave `PyExtensionOp` unimplemented and disable the
pauli prop tests that depended on it.

https://github.com/CQCL/tket2/blob/473c9f50be90a849ee6378a261e6e953b7145d51/tket2-py/src/ops.rs#L247-L251

https://github.com/CQCL/tket2/blob/473c9f50be90a849ee6378a261e6e953b7145d51/tket2-py/test/test_pauli_prop.py#L96-L100

I'll raise an issue to get them fixed them.


BREAKING CHANGE: Updated to latest hugr version (`hugr-rs 0.12` /
`hugr-py 0.8`). ECCs need to be recompiled.

---------

Co-authored-by: Douglas Wilson <141026920+doug-q@users.noreply.github.com>