ci: Fix incorrect wheels name on publish (#587)
Fixes #586, which was broken when we moved `pyproject.toml` in #546.

Running `maturin build` from the root dir tries to build the workspace
project, creating an invalid wheel name. This PR updates the ci job and
the `just build` command to run maturin from inside `/hugr-py/`.

Maturin is still configured as a build tool in the workspace so that
`maturin develop` works everywhere. In that case we don't care about the
project name.