feat: Support encoding float and sympy ops (#618)
So we can encode circuits that either encode sympy expressions, or use
the float extension to manipulate the rotations.

Note that this does not yet support `RotationOps::from_halfturns`, since
that returns an `Option<Rotation>` since unwrapping that requires
control flow operations, so the parameter encoder would need to do
constant propagation with parameters across control flow...

drive-by: Mention that `RotationOps::from_halfturns` is fallible in its
description. This required a non-breaking update to the extension's
`json` definition.