feat: Decode pytket op parameters (#644)
This is an alternative to #635, using pure-rust instead of calling to
python so we can keep the pytket decoder on the `tket2` crate.

Improves parsing of pytket operation parameters by defining a grammar
and parser for sympy expressions using `pest` (based on the [calculator
example](https://pest.rs/book/examples/calculator.html) for infix
operation precedence).

- Unrecognized operations are still put inside an opaque `SympyOp`, but
that should be easy to change in the future.
- I tested multiple sympy expressions to ensure we are able to parse
them, but unrecognized ones will also fallback to `SympyOp`.

This is still missing routing parameters to hugr inputs (#628), as it is
blocked by https://github.com/CQCL/hugr/issues/1562.

drive-by: Move the pytket parameter encoding/decoding routines to
`::serialize::pytket::param::{de,en}code`.

Closes #637.