feat!: Track circuit extensions and read/write packages (#680)
Removes the guppy-specific and adds supports for loading functions
packages and standalone hugrs.

Temporarily keeps track of the required extensions for the hugr in an
optional `Circuit::required_extensions` field until
https://github.com/CQCL/hugr/issues/1613 gets implemented. Fallbacks to
a default set when loading bare hugrs.

Note that storing a circuit with a non-root parent is currently an
error. We'll need to store some pointer to the entrypoint on the hugr's
metadata, and that'll require some serialization-stable path encoding.
I'll open an issue for that .

blocked-by: https://github.com/CQCL/hugr/pull/1621. I'll remove the
patch in cargo.toml once that gets released.

drive-by: Use `circuit_hash` for the `PartialEq` implementation of
circuits. The derived equality failed on graphs with different node
indices.

BREAKING CHANGE: Removed `load_guppy_*` methods. Use
`Circuit::load_function_reader` instead.