feat(tket2-hseries)!: insert RuntimeBarrier across qubits in a Barrier (#866)
Closes #853
Closes #865



~Blocked by hugr release including
[InsertCut](https://github.com/CQCL/hugr/pull/2153)~

Barriers exist across arbitrary types.
To make sure scheduling respects this at runtime, when lowering we
follow it with a `RuntimeBarrier` across a generically sized array of
qubits.


To do this the lowering unpacks qubits from the barrier type row,
applies the runtime barrier, then repacks them in to the original types.

Arrays and tuples containing qubits are recursively unpacked.

Sums of qubits are ignored except in the special case of arrays of
option of qubits, wherein it is assumed all qubits are present (when
generated by guppy this is guaranteed).


For each concrete container type a temporary extension is used to add
unpacking and re-packing operations. These are later lowered to calls to
functions that implement the operations.


BREAKING CHANGE: `QSystemOpBuilder` gained supertrait `ArrayOpBuilder`

---------

Co-authored-by: Douglas Wilson <douglas.wilson@quantinuum.com>
Co-authored-by: Mark Koch <mark.koch@quantinuum.com>
Co-authored-by: Agustín Borgna <agustin.borgna@quantinuum.com>