feat(pytket-encoding): Bit initialization and bool operations (#907)
Adds a pytket emitter for the `tket2.bool` extension, and constant bits
initialization (for both `tket2.bool` and `[]+[]` sum bools).

While encoding `tket2.bool` ops is not urgent, the main goal was to
define the `make_tk1_classical_operation` and
`make_tk1_classical_expression` helpers.

Unit tests are again waiting for the decoding support, so we can test
the roundtrips.

---------

Co-authored-by: Alec Edgington <54802828+cqc-alec@users.noreply.github.com>