fix: Use `TypeArg::List` instead of `TypeArg::Tuple` (#944)
This replaces `TypeArg::Tuple` with `TypeArg::List` in the construction
of a function type. It also changes the `Cargo.toml` to use the main
branch of `hugr` so that https://github.com/CQCL/hugr/pull/2378 is
integrated, fixing more test failures.

There's now one remaining test failure, related to
https://github.com/CQCL/hugr/issues/2387. It's unclear to me whether
that is to be fixed in `tket2` or in `hugr`.