feat: Add tket1-passes subcrate (#939)
This introduces a way to call into TKET1's C++ library from here
(tket2). In writing this, it was my priority to

1. not complicate the build process for anyone that is not interested in
this functionality, and
2. not slow down CI too much.

By not including `tket1-passes` in Cargo's `default-members`, this will
not be compiled unless required explicitly. One way to simplify the
compilation process further would be to provide pre-compiled libraries
for the main platforms. This requires a certain maintenance burden on
us, however, which is why I haven't gone down this route (yet).

On CI speed: this uses a pre-compiled version of tket1, so that the
compilation times for the library are low (~30s). The library is
uploaded as artifact and re-used for every job. I've also amended the CI
job `rust-stable-no-features` to not compile `tket1-passes` (as well as
the other CLI tools that have no tests) for further speed up. A way to
speed up the CI further would be to cache the compiled library across
workflow runs. I'll create an issue for this.

Finally, a note on platform support: the current build scripts will
probably not work on Windows. I am not familliar with Windows tooling
and had no way of testing it. Is this something that must be tackled
before merging @aborgna-q ?

Closes #940

---------

Co-authored-by: Agustín Borgna <121866228+aborgna-q@users.noreply.github.com>
Co-authored-by: Alec Edgington <54802828+cqc-alec@users.noreply.github.com>