chore: Fetch tket-c-api from conan remote (#1002)
This greatly simplifies the build! No setup is required anymore. Just
`uv run cargo build` and you are good to go -- or `cargo build` with
`conan` in your path. In addition, no C++ compilation should happen if
you are on one of the supported platforms (Linux X86, Mac ARM, Mac X86
or Windows X86).

As before, If you do not want to use `conan`, you just need to set
`TKET_C_API_PATH` to point to the install folder of a compiled dynamic
library of `tket-c-api`.