test: Add some example guppy-generated hugrs (#904)
Adds guppy program examples described by @cqc-alec in
https://github.com/CQCL/hugr/issues/2246.

These are _example programs_ intended to test compiler passes
internally. We are not interested in testing _guppy_ here but rather
programs that resemble guppy's compilation output in their structure,
hence why we use a pinned guppy version and don't enforce the artifacts
to be up-to-date.

We also generate `.mmd` mermaid diagrams if the example defines a
`.generate_mmd` marker file.

---------

Co-authored-by: Alan Lawrence <alan.lawrence@cambridgequantum.com>