feat: Add ResourceScope (#1052)
Alright, couldn't resist it. Here's a PR that moves a lot of the logic
that I tried to put into `portgraph` into `tket`. I think it makes much
more sense this way.

I'm interested to see what you think! (a lot of the diff is docs and
snapshot tests, so don't worry :P)