feat: Add a `borrow_array` type replacement pass (#975)
Draft because performance benchmarks with Guppy using this are needed
first to see if this is suitable. There are probably a lot of ways that
some of the more repetitive code in all the `dest` functions could be
generalised but for now I prioritised getting something working.

Closes [#2397](https://github.com/CQCL/hugr/issues/2397) (but should add
a new issue for doing this properly in LLVM at some point)