refactor!: Factor out wasm extension code into compute module (#1089)
## Changes:
- Add a `compute` module which contains `ComputeOp` and `ComputeType`
which are parameterised by the extension that they're being used for.
- Add a `extension::wasm::WasmExtension` struct to parameterise those
structs.
- Make macros for defining the `OpDef` versions of ops and defining
traits on them, and for the builder interface. The wasm module now
mostly calls these macros.
- Breaking: Split the `add_const_module` from the `WasmOpBuilder`
interface, this needs to be different for a const gpu, it either needs
to be separated or generalised, which is a breaking change either way.
- Breaking(?): Negligably update the descriptions of ops in the wasm
extensions


BREAKING CHANGE: Move `add_const_module` from `WasmOpBuilder` into new
trait, `ConstWasmBuilder`. Update descriptions of wasm ops json.