ci: Test minimum dependencies in the release checks (#1104)
In addition to testing `tket-py` with the latest published dependencies
before releasing it,
it also checks that the _minimum_ declared dependencies also work as
recommended by the [uv
docs](https://docs.astral.sh/uv/concepts/resolution/#resolution-strategy).

This is specially important now that we plan to support multiple
versions of `tket-exts` for backwards compat.

[~Test run~](https://github.com/CQCL/tket2/actions/runs/17677623613)
[Test run](https://github.com/CQCL/tket2/actions/runs/17678423897)