test: Optimize and validate guppy hugr examples (#1116)
Add test pass runs over the circuits from #904.

Some notable issues:
- The constant folding pass (and DCE) fails on some cases, leaving
unconnected qubit ports.
  https://github.com/CQCL/hugr/issues/2557
  Should be fixed on https://github.com/CQCL/hugr/pull/2560
  - (Requires Hugr `0.22.4` / `0.23` release)
- The basic-block merge rewrite should define a composable pass.
  https://github.com/CQCL/hugr/issues/2556
- We are missing a simple `InlineCFG` pass to unwrap simple linear CFGs.
(Either single blocks, or it can be conbined with `merge_bbs`).
  https://github.com/CQCL/hugr/issues/945

If we fix those, we should be able to `assert!` that some of the
examples here optimize to the identity.
We'll need to call pytket optimisations for the rest.

- ~Requires a `hugr 0.22.3` release with this fix:
https://github.com/CQCL/hugr/pull/2549~
- Requires #1118