feat: pull out unpack functionality from barrier handling (#1144)
The qsystem lowering in #865 introduced a load of lazy type unpacking
logic that may be generally useful elsewhere.

This PR pulls out the recursive, cached:
- Analysis of what row a container type like array of tuple will be
unpacked to when extracting elements of a given type (`type_unpack.rs`)
- Building of the unpack and re-pack functions (`unpack_container.rs`)


It also extracts the map from temporary extension op to function
definition, as this trick may also be generally useful for deferred
lowering/linking (`op_function_map.rs`)


Both caches use RefCells for interior mutability, though in the map case
it is not strictly a cache - the map is a necessary output of the logic.