feat: Deal with non-local edges in pytket encoder (#1145)
When encoding a hugr into a pytket circuit, we use a toposort over the
given region to process nodes.
This skips over non-local edges, and caused the encoding to fail due to
the input wires not being registered.

This PR adds a check before processing each node that looks at any
incoming non-local edge and
- If it is a module-level definition of a float/rotation parameter,
loads it.
- If it is a module-level function definition/declaration, we know
function-typed wires don't carry pytket values so we mark them as such.
- Anything else (e.g. a non-local edge from a parent DF region) causes
the node itself to be marked as unsupported.