feat: Definition of extension ops for modifiers and global phases (#1137)
This PR defines new ExtensionOps for modifiers (control/dagger/power)
and global phases in a new module `rich_circuit`.
Modifiers are defined as higher order functions that takes a circuit and
returns the modified circuit.
Global phases are defined as an Extension op that takes a rotation angle
and returns nothing.
The resolver (Hugr to Hugr function that eliminates those) are to be
implemented in this module.

# TODOs
Three todos are left in the code:
- I haven't yet understood what `post_opdef` should do, so I left them
as default. But I am not sure this is right.
- The current `contain_qubit_term` function just cares `Runtime`,
`ListType` and `TupleType`, which might not be exhaustive enough. This
todo comment is left as it may be implemented in the future.