fix(encoded-circ): Track unsupported wires between input and output (#1224)
Depends on #1211

These are not represented in the pytket circuit, and cannot be encoded
into a SiblingSubgraph.

We just track them as an additional item in the `EncodedCircuit`'s
`EncodedCircuitInfo`.

Note that the info is only store for circuits we encode directly, and
not for nested regions inside circuit boxes.
This means that the info is lost when encoding those.
I added a commented-out test with a TODO to fix that (we'll need some
extra plumbing to match circ boxes to external metadata).

drive-by: Make sure the encoder's WireTracker stores the input parameter
names, and passes it along.

I'm ignoring breaking changes since they only affect unpublished code.