fix: Multiple fixes to the pytket encoder (#1226)
This PR is a compilation of fixes to the pytket encoder/decoder. With
these implemented, the example from
https://github.com/CQCL/hugr/pull/2667 should work correctly.

The fixes include (one per commit):

- Marking wires that carry no value as unsupported, since they cannot be
attached to a pytket operation. (They will instead be included in an
opaque subgraph / barrier op).
- Don't use `SiblingSubgraph` to define the unsupported subgraphs. That
structure is quite opinionated and brittle;
  - it doesn't support non-value edges in the boundaries
- it has a complex definition for mapping between boundary ports and
actual node ports (unnecessary in this case)
- requires running an expensive convexity checker on each instantiation,
even when we know the set of nodes is valid
  - and other annoyances
- Mark any node with order edges as unsupported (so we don't lose those
the edges the pytket roundtrip)
- Fix issue with bits not appearing in the input/output missing from the
pytket circuit's list
- Add an initial hardcoded translator between native booleans and tket
booleans, with a TODO to make the system extensible.
- Avoid `QAlloc`ing qubits that don't get consumed afterwards.
- Ensure we give loaded output parameters the correct type.
- Don't try to encode wires with complex types that include parameters
like `[float]` / `[float, rotation]`.
- Store input parameters to `extra_subgraph` (that cannot be encoded in
the pytket circuit)

The main result is a bunch of tests now passing in `tests.rs`.