fix: Don't use opgroup in pytket barrier encoding (#1251)
We were writing `OPAQUE_HUGR` to the pytket barrier operation's opgroup
when encoding opaque subgraphs, to be able to quickly check if an
incoming barrier should be decoded and return errors on invalid
payloads.

However, this was an invalid use of the `opgroup` field which caused
pytket to panic when multiple operations with the same opgroup had
different signatures.

This PR removes the use of that field, and adds some helpers to check
the contents of a barrier payload instead.