ci: Add missing tket-c-api installs (#1254)
When the TKET_C_API_PATH env var is set, `tket1-passes` looks for the
shared library in the given path rather than trying to build it by
itself.
We use the to be able to cache a single artifact and share it around.

Although the env var is set globally, we weren't calling the install
action in a couple jobs.

This should fix the errors seen in #1250