fix(pytket-decoder): Avoid QAllocating and immediately freeing qubits (#1256)
When decoding a pytket circuit, when a qubit register is not present in
the region's inputs we delay `QAlloc`ating it until it is consumed.

On the other side, at the end of the decoding we insert `QFree`s on each
qubit that didn't get consumed and is not part of the region's outputs.

When a qubit falls in both categories (i.e. it didn't get used on pytket
commands and didn't appear in the input/outputs) we were allocating it
when requesting the wires to QFree.

This PR adds a simple check to avoid the extra operations.

---------

Co-authored-by: Alec Edgington <54802828+cqc-alec@users.noreply.github.com>