fix: Use qsystem extensions on Tk2Circuit.from_bytes/str (#1296)
The loading of hugrs in the python bindings did only use the hugr+tket
extension set, without the qsystem extensions. This PR is a quick-fix to
that.

The call has an optional parameter to find a named function in the Hugr
and set it as entrypoint. In that case, we cannot currently pass a
custom extension registry.
This parameter is a vestigial option from before hugrs had entrypoints
themselves that will get removed with the Circuit removal (#1060), so
I'm not fixing that here.